Restore verification integrity after PR 585
Repair the eight semantic-policy movements, harvest the sixteen resolved debt rows, and bind accepted equivalences to address, contract, machine status, and evidence. Make standalone semantic checks reject stale reccmp reports.

Make stats baseline updates fail closed on unreviewed per-function regressions and unpairing, record exact reviewed score transitions, and bind progress snapshots to a source-model fingerprint.

Restore the union-free TUiStyleBytes allocation/reset cluster to exact matching, correct the affected Windows ABIs and runtime harness model names, and fix source/lint defects exposed by the full gate.

Verified with just precommit (including 10/10 PR runtime scenarios) and just agent-check.

+++ src/game/city_ui/TCountry.cpp
@@ -239,8 +239,8 @@ void TCountry::SetSerializedField8c(short value) {
 // FUNCTION: IMPERIALISM 0x004d7170
 short TCountry::GetOrComputeOverlayAnchorTileIndex() {
   if (overlayAnchorTileCache8c == -1) {
-    overlayAnchorTileCache8c =
-        g_pGlobalMapState->ComputeRepresentativeTileIndexForNationWithWrapBias(nationSlot, 1);
+    overlayAnchorTileCache8c = static_cast<short>(
+        g_pGlobalMapState->ComputeRepresentativeTileIndexForNationWithWrapBias(nationSlot, 1));
   }
   return static_cast<short>(overlayAnchorTileCache8c);
 }
